backend/requirements.txt
flask==2.3.3
groq==0.11.0
requests==2.31.0
python-dotenv==1.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
flask-login==0.6.3
bcrypt==4.2.0
httpx==0.27.2